+++ CHANGELOG.md
@@ -32,6 +32,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed bugs
 
+* Revsets now support `\`-escapes in string literal.
+
+
 ## [0.16.0] - 2024-04-03
 
 ### Deprecations